+++ script/parser/tokens.lua
@@ -1,6 +1,6 @@
 local m = require 'lpeglabel'
 
-local Sp     = m.S' \t'
+local Sp     = m.S' \t\v\f'
 local Nl     = m.P'\r\n' + m.S'\r\n'
 local Number = m.R'09'^1
 local Word   = m.R('AZ', 'az', '__', '\x80\xff') * m.R('AZ', 'az', '09', '__', '\x80\xff')^0