+++ changelog.md
@@ -5,6 +5,7 @@
 * `CHG` `document-symbol`: redesigned to better support for `Sticky Scroll` feature of VSCode
 * `FIX` `diagnostics.workspaceDelay` can not prevent first workspace diagnostic
 * `FIX` [#1476](https://github.com/sumneko/lua-language-server/issues/1476)
+* `FIX` [#1493](https://github.com/sumneko/lua-language-server/issues/1493)
 
 ## 3.5.3
 `2022-8-13`