+++ changelog.md
@@ -7,6 +7,7 @@
 * `FIX` [#1476](https://github.com/sumneko/lua-language-server/issues/1476)
 * `FIX` [#1493](https://github.com/sumneko/lua-language-server/issues/1493)
 * `FIX` [#1499](https://github.com/sumneko/lua-language-server/issues/1499)
+* `FIX` [#1526](https://github.com/sumneko/lua-language-server/issues/1526)
 
 ## 3.5.3
 `2022-8-13`