+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 3.5.4
+`2022-9-6`
 * `NEW` `type-formatting`: fix wrong indentation of VSCode
 * `CHG` `document-symbol`: redesigned to better support for `Sticky Scroll` feature of VSCode
 * `FIX` `diagnostics.workspaceDelay` can not prevent first workspace diagnostic